requirements.txt
beautifulsoup4==4.11.1
cloudscraper==1.2.64
img2pdf==0.4.4
pandas==1.5.0
Pillow==9.2.0
pygame==2.1.2
requests==2.22.0
